skipping tuple tests when C* version < 2.1

+++ tests/integration/cqlengine/columns/test_container_columns.py
@@ -27,6 +27,7 @@ from cassandra.cqlengine.models import Model, ValidationError
 from cassandra.cqlengine.management import sync_table, drop_table
 from tests.integration.cqlengine import is_prepend_reversed
 from tests.integration.cqlengine.base import BaseCassEngTestCase
+from tests.integration import greaterthancass20, CASSANDRA_VERSION
 
 log = logging.getLogger(__name__)
 
@@ -547,12 +548,15 @@ class TestTupleModel(Model):
     mixed_tuple = columns.Tuple(columns.Text, columns.Integer, columns.Text, required=False)
 
 
+@greaterthancass20
 class TestTupleColumn(BaseCassEngTestCase):
 
     @classmethod
     def setUpClass(cls):
-        drop_table(TestTupleModel)
-        sync_table(TestTupleModel)
+        # Skip annotations don't seem to skip class level teradown and setup methods
+        if(CASSANDRA_VERSION >= '2.1'):
+            drop_table(TestTupleModel)
+            sync_table(TestTupleModel)
 
     @classmethod
     def tearDownClass(cls):
@@ -744,12 +748,15 @@ class TestNestedModel(Model):
     set_tuple = columns.Set(columns.Tuple(columns.Integer, columns.Integer), required=False)
 
 
+@greaterthancass20
 class TestNestedType(BaseCassEngTestCase):
 
     @classmethod
     def setUpClass(cls):
-        drop_table(TestNestedModel)
-        sync_table(TestNestedModel)
+        # Skip annotations don't seem to skip class level teradown and setup methods
+        if(CASSANDRA_VERSION >= '2.1'):
+            drop_table(TestNestedModel)
+            sync_table(TestNestedModel)
 
     @classmethod
     def tearDownClass(cls):
